backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
drf-spectacular
django-environ
cloudinary
django-cloudinary-storage
Pillow
reportlab
django-filter
